Merge in EDELIVERY/domibus from bugfix/EDELIVERY-9462-500-error-when-super-user-disables-default-domain-after-switching-preferred to development

* commit 'b211fee7ce2e385bbf33e0a4c0b7a86a9ea4db65':
  [EDELIVERY-9462] do not allow the superuser to disable its current domain (client-side)
  [EDELIVERY-9462] fix unit test
  [EDELIVERY-9462] do not allow the superuser to disable its current domain

+++ Domibus-MSH/src/main/java/eu/domibus/web/rest/DomainResource.java
@@ -1,6 +1,7 @@
 package eu.domibus.web.rest;
 
 import eu.domibus.api.multitenancy.Domain;
+import eu.domibus.api.multitenancy.DomainContextProvider;
 import eu.domibus.api.multitenancy.DomainService;
 import eu.domibus.api.multitenancy.UserDomainService;
 import eu.domibus.api.security.DomibusUserDetails;
@@ -40,14 +41,17 @@ public class DomainResource {
 
     private final UserDomainService userDomainService;
 
+    private final DomainContextProvider domainContextProvider;
+
     public DomainResource(DynamicDomainManagementService dynamicDomainManagementService,
                           DomainService domainService, DomibusCoreMapper coreMapper,
-                          AuthenticationService authenticationService, UserDomainService userDomainService) {
+                          AuthenticationService authenticationService, UserDomainService userDomainService, DomainContextProvider domainContextProvider) {
         this.dynamicDomainManagementService = dynamicDomainManagementService;
         this.domainService = domainService;
         this.coreMapper = coreMapper;
         this.authenticationService = authenticationService;
         this.userDomainService = userDomainService;
+        this.domainContextProvider = domainContextProvider;
     }
 
     /**
@@ -78,6 +82,11 @@ public class DomainResource {
 
         if (StringUtils.equalsIgnoreCase(domainCode, preferredDomainCode)) {
             throw new ValidationException("Cannot disable the domain of the current user");
+        }
+
+        Domain currentDomain = domainContextProvider.getCurrentDomain();
+        if (StringUtils.equalsIgnoreCase(domainCode, currentDomain.getCode())) {
+            throw new ValidationException("Cannot disable the current domain of your session");
         }
 
         dynamicDomainManagementService.removeDomain(domainCode, true);
